pyproject.toml
[tool.poetry]
name = "avdev-sandbox"
version = "0.1.0-alpha.1"
description = "Sandbox for development and analysis of autonomous vehicles"
authors = ["Spencer Hallyburton <spencer.hallyburton@duke.edu>"]
readme = "README.md"
license = "MIT"
homepage = "https://avstack.org"
repository = "https://github.com/avstack-lab/avdev-sandbox"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
]
keywords = ["AVstack", "autonomy", "autonomous vehicles"]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "~3.10"
lib-avstack-api = { path = "submodules/lib-avstack-api", develop = true }
lib-avstack-core = { path = "submodules/lib-avstack-core", extras = ["percep"], develop = true }

[tool.poetry.dev-dependencies]
gdown = "^4.6.4"
ipywidgets = "^8.0"
jupyter = "^1.0.0"

[tool.poetry.group.test.dependencies]
snakeviz="^2.1.1"
pytest = "^7"
pytest-cov = "^4"
bandit = "^1.7"
black = "~22.12.0"
mypy = "^0"
isort = "^5.11"
flake8 = "^5.0"
autoflake = "^1"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q -s"
testpaths = [
    "tests",
    "integration",
]

[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"
